+++ skuratov.danil/S5/treeTraversalCmds.cpp
@@ -1,10 +1,10 @@
 #include "treeTraversalCmds.hpp"
 
-std::pair< size_t, std::string > skuratov::isAscending(UBST< size_t, std::string >&)
+std::pair< int, std::string > skuratov::isAscending(UBST< int, std::string >&)
 {}
 
-std::pair< size_t, std::string > skuratov::isDescending(UBST< size_t, std::string >&)
+std::pair< int, std::string > skuratov::isDescending(UBST< int, std::string >&)
 {}
 
-std::pair< size_t, std::string > skuratov::isBreadth(UBST< size_t, std::string >&)
+std::pair< int, std::string > skuratov::isBreadth(UBST< int, std::string >&)
 {}
